Refactored netlink API into link and IP modules
